restructuring libs to be a bit more sane in terms of schema and so that we can split between aws and gcp
