+++ libs/aws/aws_lambda.py
