+++ libs/aws/brute.py
@@ -16,7 +16,7 @@ import pprint
 import sys
 
 
-from libs.sql import *
+from libs.aws.sql import *
 
 pp = pprint.PrettyPrinter(indent=5, width=80)
