+++ libs/aws/ce.py
