+++ libs/aws/cloudtrail.py
