+++ libs/aws/dynamodbstreams.py
