+++ libs/aws/sqs.py
