+++ libs/aws/sts.py
