+++ libs/aws/translate.py
