+++ modules/aws/ce.py
@@ -4,7 +4,7 @@ usually have to be root or be specifically assigned the
 permission to get anything from this
 '''
 
-from libs.ce import *
+from libs.aws.ce import *
 
 
 def module_costexplorer_get_cost_and_usage():
